* Refactoring (introduce workers) * Performance fixing * Server statistics

+++ Sources/EntityFX.Gdcame.Manager/Mappers/IncrementorContractMapper.cs
@@ -11,7 +11,7 @@ namespace EntityFX.Gdcame.Manager.Mappers
         {
             return new Incrementor
             {
-                IncrementorType = (IncrementorTypeEnum) ((int) source.IncrementorType),
+                IncrementorType = (IncrementorTypeEnum)  source.IncrementorType,
                 Value = source.Value
             };
         }
